sac58r: Rayleigh: Add basic support for Rayleigh sac58r chipset/board
This patch adds basic support for Rayleigh sac58r SoC.
As the RM is still in draft and we are still missing the board and schematics,
these files will have to be updated later with correct IO muxing and configurations.

The board file integrates basic PALLADIUM support, that is still under development.

Updated for new build system.

Signed-off-by: Irina Presa <Irina.Presa@freescale.com>