FSL ESDHC driver: add delay when reading data in esdhc_send_cmd()
For some reason, we get stuck here on sac58r if we do not add
this delay. This is an ugly workaround for now, that needs to
be investigated and removed eventually.

Signed-off-by: Gilles Talis <gilles.talis@freescale.com>

Cherry-picked from "9e5df50 FSL ESDHC driver: add delay when reading
data in esdhc_send_cmd()"
Signed-off-by: Dan Nica <dan.nica@nxp.com>
Issue: ALB-3031