sac58r EVB: refine audio PLLs MFD and MFN parameters
At reset the MFN and MFD parameters of the audio PLLs are way too big:
MFN = 81604373
MFD = 536870875

Because of this, kernel is unable to compute the right values of these PLLs
and end up with a erroneous 1176MHz instead of the correct 1179MHz value.
(see arch/arm/mach-imx/clk-pllv3.c, clk_pllv3_av_recalc_rate()).

Here, we set the MFD and MFN registers of the audio PLLs to reasonable
values that kernel can work with:
MFN = 152
MFD = 1000

These values allow obtaining the value we are looking for:
PLL frequency = 1179648000 Hz

This value divided by 48 allows providing a 24,576MHz
input clock to the different audio IPs

Signed-off-by: Gilles Talis <gilles.talis@freescale.com>