DDR: implement DDR handshake at functional reset for EVB board
In order for the DDR to be stabilized after a functional reset, DDR handshake
feature was implemented and enabled (any reset functional event will not be
propagated to the device until DDR is moved to a safe state).

After a functional reset, before configuring the DDR, it is checked if the
handshake was done and if the DDR is in a safe state.

Signed-off-by: Ana Nedelcu <ana.nedelcu@nxp.com>
Signed-off-by: Cosmin-Stefan Stoica <cosmin.stoica@nxp.com>