U-boot configuration refactoring
CONFIG_EXTRA_ENV_SETTINGS has almost the same content
for all boards being a duplicate code.

In order to be able to move it to s32v234_common.h header
CONFIG_BOARD_EXTRA_ENV_SETTINGS is introduced. It contains
the additional specific configuration for board. If a board
has no specific configuration the variable is a void string.

Signed-off-by: Stoica Cosmin-Stefan <cosmin.stoica@nxp.com>