ALB-287 Move clock framework to SOC level
The clock settings from EVB board are similar to all other boards,
except the FVB board which is not longer supported, because the clock
settings are from bringup phase.

It is considered the EVB clock settings as model for SOC level. For the
other boards they will be customized depending on board target.

Signed-off-by: Stoica Cosmin-Stefan <cosmin.stoica@nxp.com>