ALB-297:s32v234: Correct CAN1 pinmuxing
According to the 'IO Signal Table' in the
S32V234_IO_Signal_Description_Input_Multiplexing_Table.xlsm attached to the
S32V234 reference manual S32V234RM Rev. 1.2, 10/2015:

 1. CAN_FD1 RXD can be routed on PA3 or PA5
 2. CAN_FD1 TXD can be routed on PA4

In the code it was implemented as PC11 and, respectively, PC10, which is incorrect.
We correct the settings and comments accordingly.

Based on patch by Heinz Wrobel, debugging by Han R.

Signed-off-by: Eddy Petrișor <eddy.petrisor@nxp.com>