mmu: MMU tables addr should be relative to IRAM_BASE_ADDR
MMU tables address should be relative to IRAM_BASE_ADDR
and not to CONFIG_SYS_FSL_IRAM_BASE, where:
- CONFIG_SYS_FSL_IRAM_BASE should be SRAM start
- IRAM_BASE_ADDR should be the address in SRAM where
u-boot is placed

MMU should map the entire SRAM and not only the SRAM space
reserved for u-boot. Otherwise, if IRAM_BASE_ADDR is not a
multiple of 2M, the MMU mapping will fail.

Issue: ALB-928

Signed-off-by: Larisa Grigore <Larisa.Grigore@nxp.com>