Enable DISTRO_DEFAULT on platforms that missed it before
A number of platforms had been using the distro default feature before
it was moved to Kconfig but did not enable the new Kconfig option when
it was enabled.  This caused a regression in terms of features and this
introduces breakage when more things move to Kconfig.

Signed-off-by: Tom Rini <trini@konsulko.com>

Also applied "3337e3a Enable DISTRO_DEFAULT on platforms that missed
it before" on the 's32' configuration files that missed it.
Signed-off-by: Dan Nica <dan.nica@nxp.com>
Issue: ALB-3024