clk:DEC200 Enable the clock for DEC200 by default
There is not kernel driver for DEC200, so the kernel
will disable the DEC200 clock affecting the DCU functionality
due to direct dependency to DEC200.

In order to avoid this issue, the clock for DEC200 is enabled from
u-boot and the gating for DEC200 from Kernel is disabled until
a DEC200 kernel driver is implemented.

Signed-off-by: Stoica Cosmin-Stefan <cosmin.stoica@nxp.com>