Disable the set of GPR8
The set of GPR8 to SRC_GPR8_2D_ACE_QOS_OFFSET decreases the
performance of the GPU and the bandwidth of the DDR chip.

The glmark2 score is decreased from 247 to 68 and the DDR
bandwidth is decreased from 1258 MiB/s to 273.730 MiB/s for
memcpy operations using 64MiB chunks (mbw benchmark).

Issue: ALB-1464
Signed-off-by: Ghennadi Procopciuc <Ghennadi.Procopciuc@nxp.com>