configs: define standalone load addr for s32v234
Define CONFIG_STANDALONE_LOAD_ADDR is not defined for s32v234 platform. It used a
generic address (defined in arch/arm/config.mk) - 0xc100000 which is not compatible
with the platform. Redefined to (DDR_BASE_ADDRESS + 0x100000).
CONFIG_STANDALONE_LOAD_ADDR is the address where standalone application
should be loaded and run in u-boot environment and is needed at the link
stage of the application (see example in /examples/standalone/Makefile)

Issue: ALB-2152
Signed-off-by: Constantin-dan Avram-nxf32119 <constantin-dan.avram@nxp.com>