lowlevel_init:Speed up SRAM init change DMA transfers 32 byte blocks
Change the the DMA to copy 32 byte blocks. This necessitated a change to
the linker script to 32 byte align the .bss and .data sections. This
linker script change is specific to s32 platforms. Add a new linker
script arch/arm/cpu/armv8/s32/u-boot.lds for the s32 platforms to the
s32 cpu directory and add an entry to the configuration file
include/configs/s32.h to point the main Makefile to the new linker
script. Update alignment checks for initram command to check for
32 byte alignment.

Issue: ALB-1980

Signed-off-by: Matthew Nunez <matthew.nunez@nxp.com>