s32v234: Set ENET-PLL DFS4 frequency to 50 MHz
In commit 78285ccae3f ("MFN and MFI for each DFS becomes configurable"),
MFI and MFN were mistakenly set to 2 and 0 respectively, instead of 20
and 0. This resulted in a DFS4 output frequency of 500 MHz instead of 50
MHz. In commit 76ebc92425 ("Set the clock frequency for SDHC to 50Mhz"),
MC_CGM_0_AC15_DC0[DIV] is set to 9 to ensure a 50 MHz SDHC CLK
frequency. Later, commit 8820f8bb3e ("Clock framework configuration
updates") updated the comment regarding the DFS4 frequency to reflect
the actual value of 500 MHz.

Starting with the S32V234 reference manual Rev. 3 (10/2017), a table of
allowable MFI/MFN settings is provided. According to it, the DFS4 output
frequency is not allowed to be 500 MHz for ENET-PLL. Therefore, we keep
the currently used 50 MHz SDHC CLK frequency, with a 50 MHz DFS output
and a MC_CGM_0_AC15_DC0[DIV] of zero.

Issue: ALB-2256
Signed-off-by: Stefan-Gabriel Mirea <stefan-gabriel.mirea@nxp.com>