kconfig: Fix duplicate "TARGET_S32V234EVB" choice
As a result of the merge with v2018.07, the
TARGET_S32V234EVB option is present in both
arch/arm/Kconfig and /arch/arm/cpu/armv8/s32/Kconfig.
It was removed from the former. Also, the ARCH_S32
was selecting ARM64 and CUSTOM_CMD_FLASH twice. This
was resolved as well.

Issue: ALB-3044
Signed-off-by: Dan Nica <dan.nica@nxp.com>