armv8: s32v234: Revert basic support introduced in commit "9702ec0"
The upstream commit "9702ec0 armv8: s32v234: Introduce basic support
for s32v234evb" introduces the basic s32v234evb support in a different
directory structure compared to the one we use locally. We'll keep
the one we used to have locally, since it is more mature and the
upstream support is quite outdated.

Issue: ALB-3048
Signed-off-by: Dan Nica <dan.nica@nxp.com>