configs: Make 's32v234evb_' a symlink to 's32v234evb29288_'
Our convention was to have separate configuration files for
'evb28899' and 'evb29288' and have 's32v234evb_defconfig'
be a symlink to the 'evb29288' one, which was considered to
be the default.

Since there was only 's32v234evb_defconfig' in upstream,
and this was the one used during the upgrade, the convention
had to be re-established.

Issue: ALB-3024
Signed-off-by: Dan Nica <dan.nica@nxp.com>