s32: lowlevel: Restrict disablement of SWT_4 to S32V234 only
The SWT_4 must be disabled on S32V234 because it is
enabled by default, and can be disabled because the core it
services, the M4, is online at that point.

On S32G2XX however, no watchdog is enabled by default
and SWT_4 services the A53_1 core, which may not be
online at that point. No SWT_* can be accessed until
the core it services is enabled.

Issue: ALB-3627
Signed-off-by: Dan Nica <dan.nica@nxp.com>