s32-gen1: Fixup psci support for A53 cores
Conditionally add/modify the "enable-method" property based on whether
U-Boot assumes the presence of a trusted firmware at EL3. If that is the
case, the "enable-method" will be "psci" and a corresponding "psci" node
is also added.
If U-Boot runs at EL3 (presumably no trusted firmware), fall back to the
existing spin-table values.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@nxp.com>
Issue: ALB-3704