u-boot: Fix compilation error with binutils-2.30
In aarch64 linker compiled from 'binutils-2.30', present on
Ubuntu 18.04 cross-toolchain, an extra check for relocations
on 32 bit addresses has been added. This led to the following error
on u-boot compilation:
"relocation R_AARCH64_ABS32 against `a local symbol' can not be used
when making a shared object"
The fix is to use 64 bit addresses on all immediate load operations
which involve symbols and not constants (from literal pool).

Issue: ALB-3883
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>