setexpr: Avoid a buffer overflow in do_setexpr
The code moves around 'tail' bytes to make room for the
in-place substitution. But 'tail' should be the amount of
bytes since the end of the to-be-substituted word until
the end of the string, including the NULL character at the end.
The 'tail' used to be calculated as the number of bytes until
the end of the allocated buffer, which means that if we have to
move the last 'tail' bytes to the right in order to make room
for the substitution we will write past the buffer boundary.

Issue: ALB-276
Signed-off-by: Dan Nica <dan.nica@nxp.com>