s32g2xximage: Complete fix for QSPI Flash Boot
The qspi flash boot support was not complete
due to missing 'CONFIG_FLASH_BOOT' in any Kconfig.
Until we find a proper way to deal with the dependency of
u-boot and IVT tool in regards to this particular 'CONFIG_FLASH_BOOT'
we temporarily add its definition in 'arch/Kconfig' in order to
be properly used in generating a correct IVT for QSPI boot flow.
Also, we remove 'CONFIG_FLASH_BOOT' from 'configs_whitelist'
because the IVT cannot use it from there.

Issue ALB-3927
Signed-off-by: Ciprian Costea <ciprianmarian.costea@nxp.com>