s32: Separate 'earlycon' from CONFIG_EXTRA_KERNEL_BOOT_ARGS
The EXTRA_KERNEL_BOOT_ARGS config replaced 'earlycon' in CONFIG_BOOTARGS
and mmcargs, but not in nfsbootargs, netargs and flashbootargs. This
config was not defined for Treerunner and S32 Gen1 simulator, such that
earlycon was no longer enabled by default on these platforms.

As the LINFlex_D module address should be taken from the 'stdout-path'
property in the 'chosen' device tree node, we remove the full syntax
(i.e. including the base address) and use just 'earlycon' in the kernel
command line variables.

CONFIG_EXTRA_KERNEL_BOOT_ARGS is to be used mainly in Yocto
configuration fragments and should be also included in all kernel
parameter variables from s32.h.

Issue: ALB-4340
Signed-off-by: Stefan-Gabriel Mirea <stefan-gabriel.mirea@nxp.com>