eeprom: Enable eeprom driver to work on bigger RAM address
The way the eeprom driver works is that it reads/writes
from/to between the RAM memory and EEPROM memory.

Since we can use RAM address space > 0x80000000, we extend
the eeprom capability of converting the string address
received on command line from the size of int to ulong.

Issue: ALB-4616
Issue: ALB-4244
Signed-off-by: Ciprian Costea <ciprianmarian.costea@nxp.com>