u-boot-s32: Increase ENV_OFFSET to 0x100000
The size of u-boot has increased past the 0xc0000
offset where the u-boot environment is being written.
This leads to corruption of the u-boot image when attempting
to write/modify the u-boot environment (e.g. saveenv).

The proposed solution is to increase the U-boot Environment
offset to value 0x100000.

Issue: ALB-4669
Signed-off-by: Ciprian Costea <ciprianmarian.costea@nxp.com>