s32-gen1: Tie amount of DDR probed by U-Boot to config value
There have been customer requests to increase the amount of DDR probed
by U-Boot. It is now simple to do that by modifying a single config
value, namely CONFIG_SYS_DDR_SIZE in the platform's defconfig (and
nothing in the code).

This change works for up to 2GB of DDR probed.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@nxp.com>
Issue: ALB-4765