xen: Lower amount of Dom0 RAM to 192M
Change the default amount of RAM allocated to Dom0 to 192M, in
order to fit better in the 512M space visible to Xen and u-boot.

Also change the kernel Image load address to let Xen find a contiguous
192M memory region between 0x84000000-0x90000000 for Dom0, and avoid
loading the kernel Image, in u-boot, to an address higher than first 512M
of DRAM (i.e. after 0xA0000000).

Issue: ALB-4990
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>