rtc: fall back to ->{read, write} if ->{read, write}8 are not provided
Similar to how the dm_rtc_{read,write} functions fall back to using
the {read,write}8 methods, do the opposite in the rtc_{read,write}8
functions.

This way, each driver only needs to provide either ->read8 or ->read
to make both rtc_read8() and dm_rtc_read() work - without this, a
driver that provides ->read() would most likely just duplicate the
logic here for implementing a ->read8() method in term of its ->read()
method. The same remarks of course apply to the write case.

Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Heiko Schocher <hs@denx.de>
Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>

This is a cherry-pick of:
commit 5113f0b65a6a1ce79f4befa4e99f3e8e4a1eb9b0
("rtc: fall back to ->{read, write} if ->{read, write}8 are not provided")

Issue: ALB-9554
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>