dm: soc: Introduce UCLASS_SOC for SOC ID and attribute matching
Introduce UCLASS_SOC to be used for SOC identification and attribute
matching based on the SoC ID info. This allows drivers to be provided
for SoCs to retrieve SoC identifying information and also for matching
device attributes for selecting SoC specific data.

This is useful for other device drivers that may need different
parameters or quirks enabled depending on the specific device variant in
use.

Reviewed-by: Simon Glass <sjg@chromium.org>
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>

This is a cherry-pick of:
commit 6d3b82df832908675fe4cf4c829779f3abc34999
("dm: soc: Introduce UCLASS_SOC for SOC ID and attribute matching")

Issue: ALB-10082
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>