s32: initsram: Invalidate cache after dma_clear
The cache should be invalidated if dma operations
are used.
Invalidate the caches before the dma operation because of the
following statement from S53 manual:
"DCIMVAC operations in AArch32 and DC IVAC instructions
in AArch64 perform an invalidate of the target address.
If the data is dirty within the cluster then a clean is
performed before the invalidate."
If invalidate_dcache_range is performed after the dma operation,
the dirty lines from cache will reach SRAM because invalidate
by address will also flush the content.

Issue: ALB-5476, ALB-4999
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>