s32:default environment: fix default environment when XEN_SUPPORT is enabled
Avoid defining bootcmd in two places when XEN is enabled
The issue is observed when scripts/get_default_envs.sh is executed:
two definitions of bootcmd are stored in object file's
.rodata.default_environment section, thus the script generated
environment contains both. When this environment is enabled on the
target, the latest one overwrites the corect command.

Issue: ALB-7261

Signed-off-by: Catalin Udma <catalin-dan.udma@nxp.com>