s32: Add macros for FAT and EXT sdcard partitions
Added macros MMC_PART_FAT and MMC_PART_EXT, corresponding to the FAT and
EXT partitions IDs on the sdcard (i.e. partitions 1 and 2).

The EXT partition will be used in Xen setups to load all the required
binaries instead of the FAT partition, while for the other setups,
the FAT partition will be used.

Issue: ALB-6618
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>