s32: xen: Update u-boot env to use generated boot script
Updated the u-boot environment to load the u-boot script and
execute it, as part of the "bootcmd" variable, instead of
manually setting up the u-boot env required for booting a
Xen setup.

The u-boot script does not include the final booti command,
therefore, the booting process can be stopped and the script
can be manually fetched and run, without automatically also booting
the Xen image. This can be useful when a user needs to dynamically
amend the environment or device-tree before booting.

Issue: ALB-6618
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>