s32gen1: qspi: read correct lut during BootROM settings reset
Reading the SEQID field from the Buffer Generic Configuration Register
(BFGENCR) returns the index of a LUT sequence. As it is now, the code
wrongly uses this index to access the LUT at that index, when it should
be accessing the start of the LUT sequence corresponding to the index.

This commit will provide a function which will return the LUT address
of the first LUT in a sequence based on a provided index.

Issue: ALB-6447
Signed-off-by: Bogdan-Gabriel Roman <bogdan-gabriel.roman@nxp.com>