s32gen1: dts: change FIRC clock frequency
Previously, we used the 51MHz frequency which is the maximum frequency that
the FIRC clock may physically have. Bootrom uses this setting so that the
PLL clocks wouldn't output a clock beyond limits for some components (as
it would if it used the ideal 48MHz value).

Unfortunately, we also cannot use the FM CMU (frequency meter clock monitor
unit) because of the high tolerance margins for FIRC and FXOSC (external
oscillator). Therefore, we have FIRC as a fixed clock, set at 48MHz.

Note that, unlike bootrom, we provide FXOSC as a source for PLLs.

Issue: ALB-7938

Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>