i2c: mxc: fit the driver for non-imx platforms
Remove imx-regs.h for non-imx platforms.
Remove the clock header if CONFIG_CLK is used.
Add a default definition for I2C1_BASE_ADDR if the DM is used and it
has not already been defined. This way, platforms which do not use the
DM will still be forced to define at least one I2C base address, and
the DM users will not have to (as it was before
9c31c53564a4bbb888c7ba0515eced303ac4de02).

Issue: ALB-4149
Issue: ALB-8594

Signed-off-by: Radu Pirea (NXP OSS) <radu-nicolae.pirea@oss.nxp.com>
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>