misc: s32cc_siul2_nvram: print correct revision for cpu rev2
For S32G274A, the soc revision is made up of major, minor and subminor.
The first 2 are read from SIUL2_0 MIDR1 while the last is read
from SIUL2_1 MIDR2.
For revision 1 we had major as 0000b -> Rev 1.x and minor 0000b -> Rev x.0.
For revision 2 we have major as 0001b -> Rev 2.x and minor 0001b -> Rev x.0
Because of this for we have to differentiate how we print the revision
based on the major. For the minor we print the minor as it is for rev 1 but
for rev 2 we have -1.

Issue: ALB-7375
Signed-off-by: Andrei Botila <andrei.botila@nxp.com>