fdt: pcie: Do not skip fdt fixups for Linux
In case of using the 'skip' option for pcie nodes in 'hwconfig',
the fdt fixup should be skipped only in case of U-Boot fdt and
not Linux fdt.

'skip' option only bypasses the U-Boot configuration of SerDes
and PCIe and not Linux. Otherwise, if the Linux fdt is not
configured correctly, including the necessary fixups, the PCIe
controller configuration will fail during Linux boot, when using
the 'skip' option in 'hwconfig'.

Issue: ALB-9200
Signed-off-by: Ciprian Costea <ciprianmarian.costea@nxp.com>