dm: uclass: Start seq numbers from parent's for aliased devices
If there are aliases for an uclass, set the base for the "dynamically"
allocated numbers starting from the parent's seq number, if available.

This is an alternative to other similar fixes to 'uclass_resolve_seq',
such as upstream commit:
8f7da5d98f4e ("dm: uclass: don't assign aliased seq numbers")

The difference is that in the case of using 8f7da5d98f4e, the seq is
chosen next to the highest alias. This does not work for PCIe, since
in the case of SoCs with more than one controller (e.g. S32G with two)
there are different seq numbers assigned to bridge and EP devices for the
case when one/all controllers are enabled, none/one/all EPs are connected
to the bridges. Not to mention PCIe switches.
There is no relation between the alias ID and the seq number.
8f7da5d98f4e will try to use as seq number for the devices on the first
controller a value after the highest alias, which is unpredictable.

Therefore the best solution for this case is to start the numbering from
parent's seq number, which appears to work rather well.

Issue: ALB-9211

Signed-off-by: Ionut Vicovan <Ionut.Vicovan@nxp.com>