bitops: Add BIT_32 and BIT_64 macros
When applying some MISRA rules, lots of issues are raised with BIT macro
on AARCH32, and cast on uint32_t would be required (Rule 10.3).
The macros BIT_32 and BIT_64 are then created for 32bit and 64bit.
Then the BIT macro defaults on BIT_64 on AARCH64,
and on BIT_32 on AARCH32.

This is an adaptation of:
'utils: Add BIT_32 and BIT_64 macros' from arm-trusted-firmware
repository.

Signed-off-by: Yann Gautier <yann.gautier@st.com>
Issue: ALB-8357
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>