hwconfig: fixup: Do not return from 'ofnode_append_prop' if property does not exist
'lenp' parameter from 'ofnode_get_property' gets set to an error
code if the property does not exist. In our case we would like to
create the property if it does not exist and not directly return.
Otherwise, this will lead to Serdes not being initialized
correctly and PCIe probe will fail.

Issue: ALB-9558
Signed-off-by: Ciprian Costea <ciprianmarian.costea@nxp.com>