tools: add scmi_dtb_node_change.py
This script can is used to switch from the SIUL2 drivers to the SCMI
ones. This will take each property/child of a SIUL2 node in the dtb and
copy it to the respective SCMI node. After that, it will disable the
SIUL2 driver nodes.

Issue: ALB-9426
Issue: ALB-9556
Issue: ALB-9779
Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>