core: Add back 'uclass_resolve_seq' and fix the sequencing algorithm
This commit does two things:
- reverts:
commit b5b11558bc2d ("dm: Drop uclass_resolve_seq()")

The sequencing scheme is not working correctly for PCIe without it.
In addition, function 'uclass_resolve_seq' is adjusted to the
new sequencing scheme.

- Start seq numbers from parent's for aliased devices

If there are aliases for an uclass, set the base for the "dynamically"
allocated numbers starting from the parent's seq number, if available.

This is an alternative to other similar fixes to 'uclass_resolve_seq',
such as upstream commit:
8f7da5d98f4e ("dm: uclass: don't assign aliased seq numbers")

The difference is that in the case of using 8f7da5d98f4e, the seq is
chosen next to the highest alias. This does not work for PCIe, since
in the case of SoCs with more than one controller (e.g. S32G with two)
there are different seq numbers assigned to bridge and EP devices for the
case when one/all controllers are enabled, none/one/all EPs are connected
to the bridges. Not to mention PCIe switches.
There is no relation between the alias ID and the seq number.
8f7da5d98f4e will try to use as seq number for the devices on the first
controller a value after the highest alias, which is unpredictable.

Therefore the best solution for this case is to start the numbering from
parent's seq number, which is working rather well.

Issue: ALB-9779

Signed-off-by: Ionut Vicovan <Ionut.Vicovan@nxp.com>