drivers: soc: Add S32CC driver
Added a UCLASS_SOC S32CC driver for reading the SoC-related
info from NVMEM, and exporting them through `soc_ops` functions.

The UCLASS_SOC provides a `soc_device_match()` function for
detecting the SoC we're running on based on given family/machine/
revision strings, which are matched against what the SoC driver
instance functions return, and if a match is found, a void *ptr
to custom data provided by the caller of the function is returned.

Thus, using the NVMEM mechanism from outside a driver (fdt.c and
start_m7.c) is no longer required, and we can rely on the SoC
driver to export the data needed (SoC family/machine/revision),
which now we need to interpret instead of relying on the NVMEM provider
to generate it and directly export it (e.g. cores mask). These cases
will be handled in a future commit.

Issue: ALB-10082
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>