s32cc: usdhc: mmc: Fix XFERTYP flags setting
Upstream (commit id: '4f01db814aa') attempts to replace
some #ifdefs with IS_ENABLED(). In this process, when setting
the xfertyp flags for a given command and data packet it
excludes the internal DMA enablement for the TUNING process.

This leads to an indefinite timeout in case of S32CC manual
tuning process.

Issue: ALB-10166
Signed-off-by: Ciprian Costea <ciprianmarian.costea@nxp.com>