common/Kconfig: Add DISPLAY_SOCINFO
Added CONFIG_DISPLAY_SOCINFO to be used for printing SoC-related
information if an UCLASS_SOC driver is used for the current platform.

This provides an alternative to DISPLAY_CPUINFO, in order to have
a better separation between SoC-related info and CPU-related info.

To print SoC-related info in the init sequence, the print_socinfo()
function must be implemented by the platform.

Issue: ALB-10082
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>