s32cc: Use SoC driver for SRAM size in MMU entries
Instead of hardcoding a maximum SRAM size per SoC, use the
SoC driver to match the exact SRAM size of the platform including
derivatives. This is then used in the boot sequence in the
`arch_cpu_init_dm()` function to fixup the MMU SRAM entry.

This is done from `arch_cpu_init_dm()` instead of `arch_cpu_init()`,
because at the time of exection of the latter, the DM is not available.

Also, removed the checking of SRAM size in `start_m7` command, since
an abort should be automatically triggered if SRAM boundaries are
exceeded.

Issue: ALB-10393
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>