soc: uclass: Add `get_platform_data()` to SOC_UCLASS API
Added the `get_platform_data()` function to the SOC_UCLASS API
which can optionally be implemented by a SOC_CLASS driver and
that should fill a (pre-allocated) custom struct passed as a
parameter to it with data stored in the driver instance's platdata.

This is useful for making the SoC Driver export various data, without
limiting to machine, family and revision.

Issue: ALB-10328
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>