s32cc: drivers/soc: Add `get_platform_data()` implementation
Added implementation for the get_platform_data() callback in
SOC_UCLASS API, used for exporting a custom struct with
platform-specific data to consumers.

The exported struct is of type `struct soc_s32cc_plat` and
the callers of this callback should ensure that the *data
pointer passed to the function is pre-allocated and points
towards a struct of the same type.

Issue: ALB-10328
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>