s32cc: soc: Enable MMU and caches early during boot sequence
To speed-up boot process, this commit adds an `early_mmu_init()`
function which sets up the MMU and enables caches when it is called
from `arch_cpu_init()`. The area for the early page tables is allocated
with the early malloc functionality.

At the end of the board_f init sequence, in the `clear_bss()` callback,
the tlb_fillptr is reset and the icache/dcache are disabled, to allow
the `initr_caches()` callback in board_r to fully reinitialize page
tables.

Issue: ALB-10393
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>