pcie: dw: Allow several attempts to enable R/W access to R/O registers
On S32G3 it was observed that in some situations (e.g. when both PCIe
controllers are enabled on S32G3), the first attempts to enable R/W access
to DBI R/O registers (from MISC_CONTROL_1) fail. This would lead to
configuration of the affected controller to be incomplete and subsequent
probing of some downstream devices (e.g. SSDs) fail.
Furthermore, it could lead to timing issues which can trigger a NoC fault.

This logic was already introduced in Linux. This represents a backport
on U-Boot v2020.04

Issue: ALB-10451
Signed-off-by: Ciprian Costea <ciprianmarian.costea@nxp.com>