mach-s32: Add a command to boot an M7 elf
The 'bootm7' command can be used to kick off the M7 core using an ELF
file loaded at a certain address. The command automatically relocates
the ELF sections based on the header information and starts the M7
core at a predetermined address. This address can either be a raw
address or a symbol name.

Issue: ALB-10531
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>