clk: scmi: Allow the "scmi_gate" command to use clock names
Sometimes it can be challenging to identify the ID of a clock over SCMI
without running "clk dump" first. To address this issue, a new extension
has been added that allows a clock or a group of clocks to be gated or
ungated based on a substring from their name.

Issue: ALB-10531
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>