drivers/misc: s32cc_gpr: Add PFE_EMACS_INTF_SEL handling for S32G
Added handling for misc read/write calls with the PFE_EMACS_INTF_SEL
offset. This will only be possible on S32G2/G3, since the mapping is
added in the linked list pointed to as the mappings' list head
only on S32G platforms through compatible-to-data matching mechanism.

Issue: ALB-10366
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>