phy: s32cc-serdes: select xpcs by xpcs_id
serdes_xlate translates lane id to PHY id. The PHY id is used to select
the XPCS, but is not always matching the desired XPCS. For example, serdes
mode 1 has PCIe 0 on lane 0 and XPCS0 on lane 1. Using the lane id as
XPCS id the driver changes XPCS 1 registers, not XPCS0.

lane_id_to_xpcs_id is introduced to get the actual XPCS id.

Issue: ALB-9943

Signed-off-by: Radu Pirea (NXP OSS) <radu-nicolae.pirea@oss.nxp.com>