s32cc:hwconfig_fixups: Fix subnode creation
Upstream commit "dm: core: Allow adding ofnode subnodes"
has one issue - it creates the node but it is unusable
because no properties can be added to it.
The reason is that the properties pointer is left NULL,
and all ofnode functions for setting properties rely on it
not being initialized and having at least one property already.

We fix this by initializing it and adding the "name" property.

Issue: ALB-10511

Signed-off-by: Ionut Vicovan <Ionut.Vicovan@nxp.com>